requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
dj-database-url>=2.1
django-cors-headers>=4.3
drf-spectacular>=0.27
Pillow>=10.3
python-decouple>=3.8
gunicorn>=21.2
whitenoise>=6.6
django-filter>=23.5
